backend/requirements.txt
fastapi==0.115.6
mangum==0.19.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
boto3==1.35.99
google-generativeai==0.8.3
Pillow==11.1.0
python-dotenv==1.0.1
httpx==0.28.1
